// ui-angular/projects/smarttree/src/lib/core/api/tree/model/treeModel.ts
/**
 * Tree domain objects
 * Tree domain objects
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: info@it4all.hu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TreeNode } from './treeNode';


export interface TreeModel { 
    rootNodes: Array<TreeNode>;
    selectedNodeIdentifier?: string;
}