// ui-angular/projects/smarttree/src/lib/core/api/tree/model/treeNode.ts
/**
 * Tree domain objects
 * Tree domain objects
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: info@it4all.hu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
// import { NavigationTargetRef } from './navigationTargetRef';
import { TreeNodeKind } from "./treeNodeKind";

export interface TreeNode {
    kind?: TreeNodeKind;
    identifier?: string;
    objectUri?: string;
    caption?: string;
    shortDescription?: string;
    icon?: string;
    actions?: Array<string>;
    styles?: Array<string>;
    childrenNodes: Array<TreeNode>;
    hasChildren?: boolean;
    expanded?: boolean;
    selected?: boolean;
    childrenNodesLoaded?: boolean;
    // navigationTarget?: NavigationTargetRef;
    level?: number;
}